Fail fast on bad substitution input during layering
This PS causes layering module to fail fast on malformed
``metadata.substitutions`` entry in a document by raising a
new exception called: SubstitutionMetadataFormatError
which raises a 400. In addition, the previous exception raised
by the secrets_manager module has been renamed to
UnknownSubstitutionError which now raises a 500 instead of a 400
as this exception will most likely be due to an internal server
error of some kind.

Change-Id: Idfd91a52ef9ffd8f9b1c06c6b84c3405acab6f16