[TrivialFix] Correct regex used in jsonpath_replace
The regex in jsonpath_replace is capturing paths with numeric
values inappropriately, because the characters [ and ] aren't
escaped, leading to things like foo0 result in an array trying to
be created. This has been corrected such that now an array
should only be created for foo[0] whereas for foo0 an object
should be created.

Unit tests have been added for the function.

Change-Id: Iebb4167a0bbb9c465c348c32f55e818c41aba53d