[TrivialFix] Deckhand client filter handling
Changes the two functions in the reivison manager in the client
to not re-dictionary a dictionary argument passed as a parameter.
Before this was triggering the encoding of the url passed to
the api to contain a full extra layer of dictionary instead of
flattening the filters into individual query parameters.

Change-Id: I3e3de03f30041c20c4386b63cb77c4eabeb8bbaa