Add functional test for chained substitution
This adds a functional test for chained substitution.

For this test to pass a subtle bug had to be fixed in
the code in DocumentDict wrapper. It was possible for
instances of this class to return a NEW object when
`self.data` was invoked causing
`substitution_src.data.update(data)` to fail when
attempting to update a substitution source with updated
layered data because -- if data={} at that moment in time --
the code in DocumentDict would return a NEW object {}.

This was because `self.get('data', {})` would return {}
on the RHS instead of self['data'] because self['data']
itself would be {} which would evaluate to false in terms
of truthiness resulting in the wrong object being returned...
essentially causing the update to fail. This has been
fixed.

Change-Id: I23ad0010e1d7df73e8e1a1456ba21b3e611bb0dd