Update patch set 4
Patch Set 4:

This is a fine functional test but it's actual not the best way to validate this feature. The unit tests that I have for this are more stringent because they provide random orderings of these documents.

I'm suggesting it's possible to pass this scenario by naively processing the documents in the exact order in which they're ingested. However, I think the tests I have elsewhere already ensure that fundamentally this is working.

Patch-set: 4
Reviewer: Felipe Monteiro <1011942@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0