Update patch set 8
Patch Set 8:

> It doesn't appear to - but does this change the data stored in the
 > Deckhand DB?  If so, we will need to care for a data migration.

No, it doesn't. I manually verified this, but not by running through a bunch of scenarios as is needed. There's a US out for that -- DECKHAND-102 -- which I want to pick up to finally get the DH + KS + Barbican tests in. After doing the deployment, writing the tests will be fairly trivial. Then just a matter of checking whether there are any lingering bugs.

Patch-set: 8
Label: Code-Review=0