Add url keys to response bodies that require them
This PS adds the "url" key to missing response bodies that are
mentioned in Deckhand's API ref [0]. This is not a complete implementation
as the base URL needs to be derived from a Keystone endpoint
lookup for Deckhand and cached in memory. This also doesn't add
query string parameters in the generated URIs [1]. Finally, this
also doesn't implement the Location header which references
the generated URI for a given revision tag as per the Revision Tags
API [2].

This also doesn't include prev/next pointers which will be done
in a follow up.

[0] e.g. http://deckhand.readthedocs.io/en/latest/api_ref.html#get-revisions-revision-id
[1] e.g. http://deckhand.readthedocs.io/en/latest/api_ref.html#get-revisions-revision-id
[2] http://deckhand.readthedocs.io/en/latest/api_ref.html#post-revisions-revision-id-tags-tag

Change-Id: I35eb9ae5fcb738bc6e425e1d23f4fcb1ef12f3e2