[test] Add integration test scenario for encrypting generic type
This PS adds an integration test scenario for validating that
encrypting a generic document type and using it as a substitution
source during document rendering works.

Deckhand will not submit all generic documents to be encrypted
to Barbican with a 'secret_type' of 'passphrase'. No encoding
is provided Deckhand-side (i.e. base64) because encoding is
deprecated in Barbican since it lead to strange behavior;
Barbican will figure out what to encode the payload as
automatically. For more information, see [0] and [1].

[0] https://bugs.launchpad.net/python-barbicanclient/+bug/1419166
[1] https://github.com/openstack/python-barbicanclient/blob/49505b9aec05ec42e0a809ec51935c2324d5f3a6/barbicanclient/v1/secrets.py#L252

Change-Id: I1964aa84ad07b6f310b39974f078b84a1dc84983